Separate the various notifiers into their packages
