+++ bots/notify/src/main/java/org/openjdk/skara/bots/notify/json/JsonUpdateWriter.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.skara.bots.notify;
+package org.openjdk.skara.bots.notify.json;
 
 import org.openjdk.skara.json.*;
