+++ bots/notify/src/main/java/org/openjdk/skara/bots/notify/slack/SlackUpdater.java
@@ -20,8 +20,9 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.skara.bots.notify;
+package org.openjdk.skara.bots.notify.slack;
 
+import org.openjdk.skara.bots.notify.*;
 import org.openjdk.skara.forge.HostedRepository;
 import org.openjdk.skara.forge.PullRequest;
 import org.openjdk.skara.json.*;
